Update lcbibl.msg
What is a "corbel" on line 157?  It doesn't make much sense from googling the definition for "corbel" - "In architecture, a corbel is a structural piece of stone, wood or metal jutting from a wall to carry a superincumbent weight, a type of bracket. A corbel is a solid piece of material in the wall, whereas a console is a piece applied to the structure."