please review this one too :)
I'm not sure if this character was carrying out "some order" as in he was ordered to by an unknown higher up outside the city, or if he was ordered to by one of the bosses in the city.  The way the sentence sounded I thought it might mean, "create some order", as in the city bosses hired him to do that.  Please check if this is correct, I haven't got to this guys backstory yet :)