check on this one?
line 57 {167} I'm not sure if this character is intentionally referring to a male "workaholic" or workaholics in general.  I fixed the english for a singular workaholic (possibly referring to a character in his backstory?) but please check if the phrase should be more generic. ie:
"And when such stupid workaholics, greedy for money, die from exposure, they are then burned in the incinerator and their earned money split in the office."