pounds vs gallons? this one needs looking at :)
In the US we typically measure large amounts of liquids in gallons, which makes this exchange a little confusing.  I do see a line that explains it's 2 pounds per flask on line 121 {262}, which is about a quart of water in US measurement (I looked up the conversion, 2lbs is about .2397 gallons, so just under a quart [ie. a quarter of a gallon]).
Should this be re-written to ask for 15 quarts? or 3.75 gallons (or maybe "just under 4 gallons")?
Sorry, I know its quest related so changing the number of quarts is probably off the table as that touches the script itself, but this is a pretty weird way to refer to liquids in the US.