please take a look at this one
For some reason 23 {117} doesn't make much sense when it follows on from 22 {116}.
I'm not sure where the term "leader" applies to getting the band back together, so I tried changing it to to "YOUR band" to try and emphasize that your character is taking over the performance.
This may not exactly follow though, since I'm not sure if this applies to both when you volunteer to perform as the singer, or if you get the other guy(Big Louie? I don't remember) to perform as the singer instead.
Does this match the original intent better?