this one may need some work
I'm not sure if there's anything that can be done about this, but it seems like this dialogue was written for a female character.  I ran into this about 7 minutes into part 8 of my Nevada stream edit, and was confused about why the player dialogue seemed to be for a female character, even though my character was male.
Is it possible to tell if the script has a male/female check to make sure the right dialogue is selected?
Or was this intentionally done for convenience or something?