Please check.  Grin vs War Face?
{133}{}{Grin! Show me your grin!!!}
This is one of the drill sergeants floats in Wind of War.  The one leading the workout session for the troops in the courtyard.
Show me your grin doesn't make much sense, but it could have been intended as a funny line of dialogue.
I've never heard "grin" used like that before, but I thought it might be the same as "Show me your war face." which is a common phrase used by drill sergeants in the US military.
Would "War face" work better? or was this supposed to be a reference or funny line?