please check this one too :)
Replaced "hectically" with "zealously".
Not sure if the mechanic is supposed to guard the door "hectically", which means characterized by intense agitation, excitement, confused and rapid movement. 
I changed it to "zealously", which means with great energy or enthusiasm in pursuit of a cause or objective.
Seem like a better fit?