not sure about this one
"{115}{}{Anyway, there were two brothers."... " One of them Tom was always helping the closest... or even distant."
I'm not sure exactly what he's referring to here.  Swapped closest for poorest, and distant for richest, but I'm not sure if that's equivalent to the original.
Any ideas?
Also swapped "sued" for "conned", seemed like it fit the character better, but what do you think?