this line has a buggy reference! able to fix?????
{151}{}{Look at the report on current inmates.}
When I clicked this line in the dialogue (part 18 1:01:40) it sent me to an "Error" message.
I checked the script and this line appears to be called from NodeAnswer004, and it points to NodeAnswer004b.
NodeAnswer004b points to lines 153, 154, and 152, of which only 152 exists, and the response appears to be based on a quest variable.

Are there are couple of lines missing here? Should we fill them in?