Is this ok?
I added this line to make it more obvious where to look for the next clue.  
I spent a couple of weeks trying to figure out where to find the next missing brahmin clue and ended up looking it up in the guide just because I didn't think to click on any of the fence art. 
My hope is that new players won't be as lost as I was.