please check on this one :)
{59301}{}{Holodisk with a recording you made while talking to Hal Wallace about getting a job for the freaks. Police will find it interesting.}
and:
{59501}{}{A holodisk with a recording you made while talking to Hal Wallace about getting a job for Bob Neil. Direct evidence.}

Not sure if "getting a job on freaks" translates correctly to "getting a job for the freaks", and a similar problem with the "Bob Neil" line.
Was that conversation about getting jobs for the characters? or plotting against them? or previously working with them? (ie the drug conversations with Sam Phillips)