are "well"  & "tell"  supposed to rhyme here?
{242}{}{Well, well, well! What do you have to tell?}
changed it to "tell me" because that's more common english, but if it's intended to rhyme then the original is ok.

{224}{}{Hal Wallace?! This upstanding man was involved in criminal activities? You understand that you are responsible for your words right now?}
Replace "man" with citizen as "upstanding citizen" is more common in police procedural tv shows, but "are responsible for your words" is a little more awkward.  Usually the officer asks if the person is 'willing to testify before a court', but I don't think this game has anything like that in this quest, so I can't assume the player will ever be called on to testify.  I went with "will be held to account" because it matches the original meaning and is more common, but what do you think?