Update wcprvtdi.msg
{154}{}{...And on top of that - abandoning your military unit. What was that talk about the "untarnished honor of the uniform"?}
//
I'm not familiar with what "untarnished honor of the uniform" is referring to, there's probably an equivalent phrase in english, but I can't think of it right now.

{150}{}{What would you know about honor, you desert hobo?! If it weren't for us, you all would have been slaves long ago. And what thanks do we get? These hagglers from the wastelands rip us off every time! So it's only fair to take what's ours, even by force.}
{151}{}{You did the right thing, letting it slip. Now back to the point.}
//
Line 151 is a dialogue response to line 150...but it doesn't make much sense
The reply is for selecting 151 is:
{160}{}{[Spits under your feet] You think some civilian rat can push me against the wall? Don't forget who's turf you're standing on!}
Any ideas what this is supposed to be? Kind of confusing to me.