What tipped you off?
{318}{}{Hmm. Interesting. What tipped you off? ...
This is in response to {316}{}{Yes! I want to be on your show!}
But this character had actually said {308}{}{The folks themselves step up onto "the podium". ...
in an earlier conversation node.
So either he tipped off the player himself (which would need a line to indicate that...possibly as a joke?) or this line might not reflect the original dialogue correctly.
For now I changed it to