{249}{}{How much longer?}
{249}{}{How much longer?}
This line is used as a segue between showing him (biker Phill) weapons to upgrade.  It doesn't make much sense as a response to most of the dialogue choices given, so I changed it to:
{249}{}{What else ya got?}
But I'm not sure if any of the other dialogue lines actually do correctly lead to "How much longer?"...so let's hope they don't.