"prematures"? and 3 untranslated lines
{469}{}{You really need 'em? Hmm. You proved yourself well with these prematures in Reno and deserve our respect. So be it - I will share them with you. You're good kid.}
changed "these prematures" to "those problems"
same on {1469}
I have no clue what "prematures" is supposed to be, but it's pretty clear he's referring to the new reno caravan quest, so I changed it to "problems".  Does this work? Is there something else he's referring to?
and the female dialogue has 3 untranslated lines: {1530}, {1532}, {1533}
also changed
{216}{}{It's simple. I'm actually not so sure myself, so - wanna bet?}
to 
{216}{}{It's simple. But if you're so sure I can't do it - wanna bet?}
Thoughts on this change? seems close to the intent and makes more sense.