big bug here!
Once you get the recording evidence from Stryker's base and bring it back, this dialog option is given to you:
{165}{}{I managed to get a recording of one of Stryker's meetings. He called you a stubborn fool, and plans to replace you after he seizes power.}
but the reply from Mordino is:
{166}{}{So, what's your final answer to my boss?}
and the next dialog option you're given is just "Error"
Looking at the script it looks like this is the relevant part:

procedure NodeShantazh5 begin
   gsay_reply(456, 164);
   if (obj_is_carrying_obj_pid(dude_obj, 761) >= 1) then begin
      giq_option(1, 456, 165, NodeShantazh6, 50);
   end
   if (obj_is_carrying_obj_pid(dude_obj, 705) >= 1) then begin
      giq_option(1, 456, 175, NodeShantazh, 51);
   end
   if (((global_var(311) bwand 4) != 0) > 0) then begin
      giq_option(1, 456, 170, NodeShantazh2, 51);
   end
   if (global_var(260) == 1) then begin
      giq_option(1, 456, 176, NodeShantazh3, 50);
   end
   giq_option(1, 456, 162, Node999, 50);
end

procedure NodeShantazh6 begin
   gsay_reply(456, 166);
   giq_option(1, 456, 167, NodeShantazhExit, 50);
end

NodeShantazhExit seems to work correctly, since dialog returns to normal, but NodeShantazh6 points to line 166 and 167, of which only 166 exists right now.
Were these lines missing in the original dialog? Can you check?
If they were missing, is there a way to figure out what they were supposed to be?