this probably needs more work
"Mister"/"Miss" might not be the best word to use here, but its used so often that replacing it would be time consuming.
What's the original intent? Maybe Gollum from Lord of the Rings? 
This comes across as an odd speech impediment, but I guess it's more or less consistent with Gollum.